ok core.js shouldn't have any of this flow stuff
